move task status into service
