+++ src/main/webapp/app/entities/programming-exercise/instructions/programming-exercise-instruction.component.ts
@@ -191,7 +191,7 @@ export class ProgrammingExerciseInstructionComponent implements OnChanges, OnDes
         setTimeout(() => {
             this.loadAndInsertPlantUmls();
             /*            this.setUpClickListeners();*/
-            this.setUpTaskIcons();
+            /*            this.setUpTaskIcons();*/
         }, 0);
     }
